github-issue3: weaver.ini ignored
Updated to read config options from both
manual instanciation (as was in the tests)
and options passed in from weaver.ini.

Changed the skip_method_list from a hashref
to an arrayref.  Forced due to Pod::Weaver's
use of Config::MVP either its limitations
or my inability to understand how to get
an arrayref out of it.

Updated tests to reflect these changes.